+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
-        <activity android:name=".example.ui.home.news.NewsCreationActivity"></activity>
+        <activity android:name=".example.ui.home.news.newsCreation.NewsCreationActivity"></activity>
         <activity
             android:name=".example.ui.login.LoginActivity"
             android:screenOrientation="portrait">
