--- app/src/main/java/ar/com/wolox/android/example/ui/home/news/INewsCreationView.kt
@@ -1,3 +1,0 @@
-package ar.com.wolox.android.example.ui.home.news
-
-interface INewsCreationView
