+++ app/src/main/java/ar/com/wolox/android/example/ui/home/news/newsCreation/INewsCreationView.kt
@@ -1,0 +1,3 @@
+package ar.com.wolox.android.example.ui.home.news.newsCreation
+
+interface INewsCreationView
