+++ app/src/main/java/ar/com/wolox/android/example/ui/home/news/newsCreation/NewsCreationPresenter.kt
@@ -1,4 +1,4 @@
-package ar.com.wolox.android.example.ui.home.news
+package ar.com.wolox.android.example.ui.home.news.newsCreation
 
 import ar.com.wolox.wolmo.core.presenter.BasePresenter
 import javax.inject.Inject
